Rename packages to modules
